Adjust properties

+++ Simple.Mox/Models/InstanceLXCs.cs
@@ -5,21 +5,21 @@ using System;
 public record InstanceLXCs
 {
     public int vmid { get; set; }
-    public int diskread { get; set; }
+    public long diskread { get; set; }
     public long mem { get; set; }
     public long disk { get; set; }
     public string status { get; set; }
-    public int diskwrite { get; set; }
+    public long diskwrite { get; set; }
     public long maxdisk { get; set; }
     public int cpus { get; set; }
     public string type { get; set; }
-    public int pid { get; set; }
-    public int netin { get; set; }
+    public long pid { get; set; }
+    public long netin { get; set; }
     public string name { get; set; }
-    public int uptime { get; set; }
-    public int netout { get; set; }
+    public long uptime { get; set; }
+    public long netout { get; set; }
     public float cpu { get; set; }
-    public int swap { get; set; }
+    public long swap { get; set; }
     public long maxswap { get; set; }
     public long maxmem { get; set; }
     public string tags { get; set; }
